move inferMem into its own file, extract recursive helper func
