+++ src/infer/infer-types.ts
@@ -1,0 +1,6 @@
+import * as tt from "./type-types";
+
+export type InferResult<T extends tt.Type = tt.Type> = readonly [
+  T,
+  readonly tt.Constraint[]
+];
